Check the limits on our line lists (#32)
* Pass all options to the minimization, but ignore all which
is not required by using **kwargs.

* Preparing removal of lines from "normal" line list in order
to make the optimized version for cool stars.

* Give a warning in the log file when a high Teff is found
and the normal linelist is used. If Teff is lower than 5200K
and the normal linelist is used, then remove the lines to
convert the normal linelist into the cool linelist, if and
only if the option 'teffrange' is set. This option is requested
in issue #20.

* The lines different in the cool line list and the normal line list.

* Create new option for removing lines for cool stars.

* Remove unnecessary comment.

* Fixed bugs concerning the feature for issue #20. All set now.